backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pandas==2.1.3
scikit-learn==1.3.2
djangorestframework-simplejwt==5.3.1
